fix(db): supabase db reset'i sıfırdan baştan sona çalışır hale getir
Bu ortamda supabase db reset hiçbir zaman tamamlanamıyordu — squash
edilmiş 00000000000000_base_schema.sql ile onu izleyen kronolojik
migration'lar arasında birçok tutarsızlık birikmişti. Tümü tek tek
bulunup düzeltildi (supabase start ile fresh volume üzerinde
doğrulandı, sonunda başarıyla tamamlandı):

- crowd_checkins/price_verifications/business_reviews/review_ratings/
  business_owners gibi squash'ta kaybolmuş veya yeniden adlandırılmış
  tablo/view'lara referanslar gerçek karşılıklarına (business_checkins,
  menu_item_price_votes, is_owner_of_business(), vb.) düzeltildi veya
  defansif (to_regclass) korumaya alındı.
- visits/loyalty_programs gibi tablolarda squash'ın eksik bıraktığı
  kolonlar ALTER TABLE ADD COLUMN IF NOT EXISTS ile tamamlandı.
- cube/earthdistance extension'ları hiç kurulmuyordu, eklendi.
- review_ratings tablosu hiç oluşturulmamıştı (sadece okunuyordu),
  oluşturuldu.
- CASE ifadesinin ELSE dalına RAISE EXCEPTION koymak gibi geçersiz
  PL/pgSQL sözdizimleri düzeltildi (submit_content_report_v3).
- RETURNS TABLE (OUT parametreli) bir fonksiyonda "RETURN NEXT
  row(...)" geçersizdi — get_business_badges_v1'de OUT değişkenlerine
  atama + boş RETURN NEXT'e çevrildi.
- Birden fazla migration aynı fonksiyonu (get_my_loyalty_cards_v1,
  get_business_chain_info_v1, get_staff_performance_today_v1) farklı
  dönüş tipleriyle CREATE OR REPLACE etmeye çalışıyordu — DROP FUNCTION
  IF EXISTS eklendi veya çakışan "stub" kaldırıldı.
- CREATE OR REPLACE VIEW mevcut kolon sırasını/tipini değiştiremediği
  için reviews_count'a cast, neighborhood'a sona taşıma uygulandı.
- İki migration çifti aynı zaman damgasını paylaşıyordu
  (20260708000001, 20260709000001) — ikincisi benzersiz bir zaman
  damgasına yeniden adlandırıldı.
- reviews.content yerine hiç var olmayan reviews.body'ye erişen
  trg_extract_review_tags_v1 düzeltildi — bu trigger her onaylı yorum
  INSERT'inde tetikleniyor, canlıda gerçekten kırıp kırmadığı ayrıca
  doğrulanmalı.
- reviews.quality_score hiç var olmadı; ona bağlı rozet koşulları
  (Kaliteli Yorum Mekanı, Gurme Seçimi, Sadık Topluluk) kasıtlı olarak
  hep false bırakıldı — yeni ürün mantığı icat edilmedi, mevcut "hiç
  verilmedi" durumu korundu.
- "Örnek Yeedoy" demo işletmesi artık sabit bir id kullanıyor (sonraki
  migration'lar zaten bunu varsayıyordu) ve gerekli placeholder
  auth.users kayıtları ekleniyor.

supabase start (fresh volume, --no-backup) ve supabase db reset ikisi
de artık sıfırdan başarıyla tamamlanıyor.

+++ supabase/migrations/20260620000006_businesses_with_stats_open_price.sql
@@ -22,7 +22,7 @@ select
   b.lng,
   b.is_active,
   b.created_at,
-  coalesce(r.reviews_count, 0)                         as reviews_count,
+  coalesce(r.reviews_count, 0)::integer                as reviews_count,
   (coalesce(r.avg_rating, 0::numeric))::numeric(3, 2)  as avg_rating,
   b.is_verified,
